+++ src/main/java/racingcar/application/implement/CarRacing.java
@@ -1,4 +1,4 @@
-package racingcar.racing;
+package racingcar.application.implement;
 
 import static racingcar.common.RacingCarConstant.MAX_RANDOM_BOUND;
 import static racingcar.common.RacingCarConstant.MIN_RANDOM_BOUND;
@@ -6,6 +6,7 @@ import static racingcar.common.RacingCarConstant.MOVE_THRESHOLD;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
+import racingcar.domain.CarRacer;
 
 public class CarRacing implements Racing<CarRacer> {
