+++ src/main/java/racingcar/application/implement/Racing.java
@@ -1,6 +1,7 @@
-package racingcar.racing;
+package racingcar.application.implement;
 
 import java.util.List;
+import racingcar.domain.Racer;
 
 public interface Racing<T extends Racer> {
