+++ src/main/java/racingcar/application/implement/RacingCarHistoryWriter.java
@@ -1,9 +1,10 @@
-package racingcar.racing;
+package racingcar.application.implement;
 
 import static racingcar.common.FormatConstants.LINE_BREAKER;
 
 import java.util.List;
 import java.util.stream.Collectors;
+import racingcar.domain.CarRacer;
 
 public class RacingCarHistoryWriter {
