+++ src/main/java/racingcar/common/support/ObjectConverter.java
@@ -1,4 +1,4 @@
-package racingcar.application;
+package racingcar.common.support;
 
 import java.util.List;
