+++ src/test/java/racingcar/CarRacerObjectConverterTest.java
@@ -6,8 +6,8 @@ import camp.nextstep.edu.missionutils.test.Assertions;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.application.service.RacingCarConverter;
-import racingcar.racing.CarRacer;
+import racingcar.common.support.RacingCarConverter;
+import racingcar.domain.CarRacer;
 
 class CarRacerObjectConverterTest {
