+++ src/test/java/racingcar/application/service/WinnerIdentifierTest.java
@@ -6,7 +6,8 @@ import camp.nextstep.edu.missionutils.test.Assertions;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.racing.CarRacer;
+import racingcar.application.implement.WinnerIdentifier;
+import racingcar.domain.CarRacer;
 
 class WinnerIdentifierTest {
